Refactored components for that sweet, swwet clode climate score.
