+++ coach-me/src/components/clients/onboarding/accountCreation/accountOne/accountOne.scss
@@ -13,6 +13,10 @@
         align-items: center;
         border-bottom: 2px solid dodgerblue;
         font-family: 'Poppin', sans-serif;
+        .fas {
+            font-size: 3em;
+            margin-left: 0.5em;
+        }
 
         h4 {
             font-size: 3em;
