+++ coach-me/src/components/clients/onboarding/createProfile/profileFive/ProfileFive.jsx
@@ -2,95 +2,14 @@ import React,{ useState } from 'react';
 import  { Link } from 'react-router-dom';
 
 import ProfileFour from '../profileFour/ProfileFour';
-import Timeslot from './Timeslot';
+import Weekdays from './Weekdays';
+import Weekends from './Weekends';
 import TimeslotWeekend from './TimeslotWeekend';
 
 import './profileFive.scss';
 
 
 const ProfileFive = (props) => {
-//   Created these for dummy data, i assume it will be brought in from the backend later
-const [weekdaySchedule, setWeekdaySchedule] = useState(
-    [
-        {
-            id: 0,
-            text: '6-10 am',
-            selected: false,
-        },
-        {
-            id: 1,
-            text: '10-12pm',
-            selected: false,
-        },
-        {
-            id: 2,
-            text: '12-2 pm',
-            selected: false,
-        },
-        {
-            id: 3,
-            text: '2-4 pm',
-            selected: false,
-        },
-    ]
-)
-
-const [weekendSchedule, setWeekendSchedule] = useState(
-    [
-        {
-            id: 0,
-            text: '6-10 am',
-            selected: false,
-        },
-        {
-            id: 1,
-            text: '10-12pm',
-            selected: false,
-        },
-        {
-            id: 2,
-            text: '12-2 pm',
-            selected: false,
-        },
-        {
-            id: 3,
-            text: '2-4 pm',
-            selected: false,
-        },
-    ]
-)
-    
-
-
-    const selectTimeWeekday = id => {
-    
-        setWeekdaySchedule(weekdaySchedule.map(item => {
-            console.log(item)
-          if(item.id === id) {
-           return {...item,selected : !item.selected}
-          }else {
-            return item;
-          } 
-        
-        }
-        ))
-      }
-
-      const selectTimeWeekend = id => {
-    
-        setWeekendSchedule(weekendSchedule.map(item => {
-            console.log(item)
-          if(item.id === id) {
-           return {...item,selected : !item.selected}
-          }else {
-            return item;
-          } 
-        
-        }
-        ))
-      }
-
-
     return (
         <div>
         <div className="header">
@@ -103,41 +22,10 @@ const [weekendSchedule, setWeekendSchedule] = useState(
         <div className="scheduling">
             <p className='help-text'>As part of Coach Me you'll regularly chat with your coach. Select what time(s) work best for you.</p>
             <h4 className='heading'>Weekdays</h4>
-            <div className="weekdays time-slots">
-            
-                {
-                   weekdaySchedule && weekdaySchedule.map(item => {
-                        
-                        return (
-                            <Timeslot
-                            id={item.id}
-                            text={item.text}
-                            selected={item.selected}
-                            selectTimeWeekday={selectTimeWeekday} 
-                            />
-                        )
-                    })
-                }
-            </div>
+            <Weekdays />
             <h4 className='heading'>WeekEnds</h4>
-            <div className="weekends time-slots">
-            
-            {
-                  weekendSchedule &&  weekendSchedule.map(item => {
-                        return (
-                            <TimeslotWeekend
-                            id={item.id}
-                            text={item.text}
-                            selected={item.selected}
-                            selectTimeWeekend={selectTimeWeekend} 
-                            />
-                        )
-                    })
-                }
-            </div>
+            <Weekends />
         </div>
-        {/* SCHEDULING */}
-
         <div className="timezone">
             <h4 className="heading">Time Zone</h4>
             <select className='zone-select' name="time-zone" >
@@ -150,10 +38,8 @@ const [weekendSchedule, setWeekendSchedule] = useState(
             </select>
         </div>
         {/* TIMEZONE */}
-        
         <div className="next">
             <Link to='/createProfile6'><i className="fas fa-chevron-right"></i></Link>
-        
         </div>
     </div>
     );
